Move gists/gist6445065 to reflectfind.
